SNAP-876 Upload version 1.0.1
Plugin description and installation instruction update
