+++ css/coinsnapcf7-styles.css
@@ -1,4 +1,4 @@
-.cf7-coinsnap {
+.coinsnapcf7 {
   .description {
     margin: 2px 0 5px;
     color: #646970;
@@ -13,7 +13,7 @@
     display: block;
   }
 
-  .cf7-coinsnap_row {
+  .coinsnapcf7-row {
     margin-bottom: 15px;
   }
