+++ library/Client/Server.php
@@ -1,7 +1,11 @@
 <?php
 declare(strict_types=1);
-
 namespace Coinsnap\Client;
+
+if (!defined('ABSPATH')) {
+    exit;
+}
+
 use Coinsnap\Result\ServerInfo;
 
 class Server extends AbstractClient {
@@ -14,7 +18,7 @@ class Server extends AbstractClient {
         if ($response->getStatus() === 200) {
             return new ServerInfo(json_decode($response->getBody(), true, 512, JSON_THROW_ON_ERROR));
         } else {
-            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), $response);
+            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), esc_html($response->getStatus()), esc_html($response->getBody()));
         }
     }
     
@@ -28,7 +32,7 @@ class Server extends AbstractClient {
         if ($response->getStatus() === 200) {
             return true;
         } else {
-            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), $response);
+            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), esc_html($response->getStatus()), esc_html($response->getBody()));
         }
     }
 }
