+++ library/Exception/BadRequestException.php
@@ -1,10 +1,11 @@
 <?php
-
 declare(strict_types=1);
-
 namespace Coinsnap\Exception;
 
-class BadRequestException extends RequestException
-{
+if (!defined('ABSPATH')) {
+    exit;
+}
+
+class BadRequestException extends RequestException {
     public const STATUS = 400;
 }
