+++ library/Exception/ForbiddenException.php
@@ -1,10 +1,11 @@
 <?php
-
 declare(strict_types=1);
-
 namespace Coinsnap\Exception;
 
-class ForbiddenException extends RequestException
-{
+if (!defined('ABSPATH')) {
+    exit;
+}
+
+class ForbiddenException extends RequestException {
     public const STATUS = 403;
 }
