+++ library/Exception/RequestException.php
@@ -1,16 +1,14 @@
 <?php
-
 declare(strict_types=1);
-
 namespace Coinsnap\Exception;
 
-use Coinsnap\Http\ResponseInterface;
+if (!defined('ABSPATH')) {
+    exit;
+}
 
-class RequestException extends CoinsnapException
-{
-    public function __construct(string $method, string $url, ResponseInterface $response)
-    {
-        $message = 'Error during ' . $method . ' to ' . $url . '. Got response (' . $response->getStatus() . '): ' . $response->getBody();
-        parent::__construct($message, $response->getStatus());
+class RequestException extends CSException {
+    public function __construct(string $method, string $url, int $status, string $body){
+        $message = 'Error during ' . $method . ' to ' . $url . '. Got response (' . $status . '): ' . $body;
+        parent::__construct($message, $status);
     }
 }
