+++ library/Http/WPRemoteClient.php
@@ -1,7 +1,10 @@
 <?php
 declare(strict_types=1);
-
 namespace Coinsnap\Http;
+
+if (!defined('ABSPATH')) {
+    exit;
+}
 
 use Coinsnap\Exception\ConnectException;
