+++ library/Result/ApiKey.php
@@ -1,9 +1,10 @@
 <?php
-
 declare(strict_types=1);
-
 namespace Coinsnap\Result;
 
-class ApiKey extends AbstractResult
-{
+if (!defined('ABSPATH')) {
+    exit;
+}
+
+class ApiKey extends AbstractResult {
 }
