+++ library/loader.php
@@ -1,9 +1,13 @@
 <?php
+if (!defined('ABSPATH')) {
+    exit;
+}
 spl_autoload_register(function ($className) {
     if (!defined('COINSNAP_SERVER_PATH')){
 	define( 'COINSNAP_SERVER_PATH', 'stores' );  
-	}
+    }
     $libName = 'Coinsnap';
+    
     // Abort here if we do not try to load Coinsnap namespace.
     if (strpos($className, $libName) !== 0) {
         return;
